Update display simulation
On MS drop and display reconnect, there is no watchlist compare. That
only happens on displays startup. This can be configured optionally.
This way, the load test can be configured to simulate the startup
behaviour, or mass disconnect / reconnect behaviour. Note that for
startup, with the watchlist compare, there should be significantly less
pods running the load because not all displays restart, and restart
times vary.